+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Updates and Fixes
 
 + Fixed `promptBeta()` period input.
++ Added storage capacity to `eiaStorageCap` data set for PADD1 middle and light distillates as a proxy for NYH.
 
 # RTL 0.1.8